fixing index view
