+++ app/helpers/posts_helper.rb
@@ -1,3 +1,2 @@
 module PostsHelper
-  include NewUserSessionHelper
 end
