+++ app/models/user.rb
@@ -12,4 +12,24 @@ class User < ApplicationRecord
   validates :password, presence: true, length: {minimum: 6}
 
   before_save { email.downcase! }
+
+
+  def self.new_token
+    SecureRandom.urlsafe_base64
+  end
+
+  def remember
+    self.remember_token = User.new_token
+    update_attribute(:remember_digest, User.digest(remember_token))
+  end
+
+  def authenticated?(attribute, token)
+    digest = send("#{attribute}_digest")
+    return false if digest.nil?
+    BCrypt::Password.new(digest).is_password?(token)
+  end
+  
+  def forget
+    update_attribute(:remember_digest, nil)
+  end
 end
