+++ app/views/posts/index.html.erb
@@ -8,8 +8,8 @@
       <tr>
         <td><%= post.title %></td>
         <td><%= post.body %></td>
-        <% if @user.logged_in? %>
-          <td><%= post.user_id.user.name %></td>
+        <% if logged_in? %>
+          <td><%= User.find_by(id: post.user_id).name %></td>
         <% end %>
       </tr>
     <% end %>
@@ -17,6 +17,6 @@
 </table>
 
 <br>
-<% if @user.logged_in? %>
+<% if logged_in? %>
   <%= link_to 'New Post', new_post_path, class: "btn btn-lg btn-primary" %>
 <% end %>
