+++ app/views/sessions/new.html.erb
@@ -10,8 +10,7 @@
 
       <%= f.label :password %>
       <%= f.password_field :password, class: 'form-control' %>
-      <%# <%=link_to "(forgot password)", new_password_reset_path %> %>
-
+      
       <%= f.label :remember_me, class: "checkbox inline" do %>
         <%= f.check_box :remember_me %>
         <span>Remember me on this computer</span>
